+++ android/app/src/main/java/com/storybooknative/MainApplication.java
@@ -3,7 +3,9 @@ package com.storybooknative;
 import android.app.Application;
 
 import com.facebook.react.ReactApplication;
+import com.reactnativecommunity.asyncstorage.AsyncStoragePackage;
 import com.learnium.RNDeviceInfo.RNDeviceInfo;
+import com.reactnativecommunity.netinfo.NetInfoPackage;
 import com.facebook.react.ReactNativeHost;
 import com.facebook.react.ReactPackage;
 import com.facebook.react.shell.MainReactPackage;
@@ -27,7 +29,9 @@ public class MainApplication extends Application implements ReactApplication {
         protected List<ReactPackage> getPackages() {
             return Arrays.asList(
                     new MainReactPackage(),
+                    new AsyncStoragePackage(),
                     new RNDeviceInfo(),
+                    new NetInfoPackage(),
                     new ComponentsPackage(),
                     new StorybookStubPackage(),
                     new SvgPackage(),