+++ packages/author-profile/src/author-profile.js
@@ -13,7 +13,6 @@ import { ratioTextToFloat } from "@times-components/utils";
 import AuthorProfileHead from "./author-profile-head";
 import { propTypes, defaultProps } from "./author-profile-prop-types";
 import authorProfileTrackingContext from "./author-profile-tracking-context";
-import Head from "./head";
 
 const AuthorProfile = ({
   adConfig,
@@ -111,7 +110,6 @@ const AuthorProfile = ({
 
         return (
           <Responsive>
-            <Head description={biography} name={name} />
             <ArticleList
               adConfig={adConfig}
               articleListHeader={articleListHeader}