+++ packages/interactive-wrapper/src/interactive-wrapper.js
@@ -52,20 +52,21 @@ class InteractiveWrapper extends Component {
   }
 
   onLoadEnd() {
-    this.webview.postMessage("thetimes.co.uk", "*");
+    if (this.webview) {
+      this.webview.postMessage("thetimes.co.uk", "*");
+    }
   }
 
-  // eslint-disable-next-line class-methods-use-this
   handleNavigationStateChange(data) {
     if (
       !data.url.includes("data:text/html") &&
       data.url.includes("http") &&
       !data.url.includes(editorialLambdaOrigin)
     ) {
+      // Need to handle native routing when something is clicked.
       InteractiveWrapper.openURLInBrowser(data.url);
-      return false;
+      this.webview.reload();
     }
-    return true;
   }
 
   render() {
@@ -90,11 +91,11 @@ class InteractiveWrapper extends Component {
         injectedJavaScript={scriptToInject}
         onLoadEnd={this.onLoadEnd}
         onMessage={this.onMessage}
+        onNavigationStateChange={this.handleNavigationStateChange}
         ref={ref => {
           this.webview = ref;
         }}
         scrollEnabled={false}
-        onShouldStartLoadWithRequest={this.handleNavigationStateChange}
         source={{ uri }}
         style={{ height }}
       />