+++ ios-app/CHANGELOG.md
@@ -3,6 +3,70 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.17.127](https://github.com/newsuk/times-components/compare/@times-components/ios-app@0.17.126...@times-components/ios-app@0.17.127) (2020-06-22)
+
+**Note:** Version bump only for package @times-components/ios-app
+
+
+
+
+
+## [0.17.126](https://github.com/newsuk/times-components/compare/@times-components/ios-app@0.17.125...@times-components/ios-app@0.17.126) (2020-06-22)
+
+**Note:** Version bump only for package @times-components/ios-app
+
+
+
+
+
+## [0.17.125](https://github.com/newsuk/times-components/compare/@times-components/ios-app@0.17.124...@times-components/ios-app@0.17.125) (2020-06-22)
+
+**Note:** Version bump only for package @times-components/ios-app
+
+
+
+
+
+## [0.17.124](https://github.com/newsuk/times-components/compare/@times-components/ios-app@0.17.123...@times-components/ios-app@0.17.124) (2020-06-19)
+
+**Note:** Version bump only for package @times-components/ios-app
+
+
+
+
+
+## [0.17.123](https://github.com/newsuk/times-components/compare/@times-components/ios-app@0.17.122...@times-components/ios-app@0.17.123) (2020-06-19)
+
+**Note:** Version bump only for package @times-components/ios-app
+
+
+
+
+
+## [0.17.122](https://github.com/newsuk/times-components/compare/@times-components/ios-app@0.17.121...@times-components/ios-app@0.17.122) (2020-06-18)
+
+**Note:** Version bump only for package @times-components/ios-app
+
+
+
+
+
+## [0.17.121](https://github.com/newsuk/times-components/compare/@times-components/ios-app@0.17.120...@times-components/ios-app@0.17.121) (2020-06-17)
+
+**Note:** Version bump only for package @times-components/ios-app
+
+
+
+
+
+## [0.17.120](https://github.com/newsuk/times-components/compare/@times-components/ios-app@0.17.119...@times-components/ios-app@0.17.120) (2020-06-17)
+
+**Note:** Version bump only for package @times-components/ios-app
+
+
+
+
+
 ## [0.17.119](https://github.com/newsuk/times-components/compare/@times-components/ios-app@0.17.118...@times-components/ios-app@0.17.119) (2020-06-12)
 
 **Note:** Version bump only for package @times-components/ios-app