+++ packages/article-comments/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.6.37](https://github.com/newsuk/times-components/compare/@times-components/article-comments@0.6.36...@times-components/article-comments@0.6.37) (2020-06-22)
+
+**Note:** Version bump only for package @times-components/article-comments
+
+
+
+
+
 ## [0.6.36](https://github.com/newsuk/times-components/compare/@times-components/article-comments@0.6.35...@times-components/article-comments@0.6.36) (2020-06-12)
 
 **Note:** Version bump only for package @times-components/article-comments