+++ packages/article/CHANGELOG.md
@@ -3,6 +3,62 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [7.4.112](https://github.com/newsuk/times-components/compare/@times-components/article@7.4.111...@times-components/article@7.4.112) (2020-06-22)
+
+**Note:** Version bump only for package @times-components/article
+
+
+
+
+
+## [7.4.111](https://github.com/newsuk/times-components/compare/@times-components/article@7.4.110...@times-components/article@7.4.111) (2020-06-22)
+
+**Note:** Version bump only for package @times-components/article
+
+
+
+
+
+## [7.4.110](https://github.com/newsuk/times-components/compare/@times-components/article@7.4.109...@times-components/article@7.4.110) (2020-06-19)
+
+**Note:** Version bump only for package @times-components/article
+
+
+
+
+
+## [7.4.109](https://github.com/newsuk/times-components/compare/@times-components/article@7.4.108...@times-components/article@7.4.109) (2020-06-19)
+
+**Note:** Version bump only for package @times-components/article
+
+
+
+
+
+## [7.4.108](https://github.com/newsuk/times-components/compare/@times-components/article@7.4.107...@times-components/article@7.4.108) (2020-06-18)
+
+**Note:** Version bump only for package @times-components/article
+
+
+
+
+
+## [7.4.107](https://github.com/newsuk/times-components/compare/@times-components/article@7.4.106...@times-components/article@7.4.107) (2020-06-17)
+
+**Note:** Version bump only for package @times-components/article
+
+
+
+
+
+## [7.4.106](https://github.com/newsuk/times-components/compare/@times-components/article@7.4.105...@times-components/article@7.4.106) (2020-06-17)
+
+**Note:** Version bump only for package @times-components/article
+
+
+
+
+
 ## [7.4.105](https://github.com/newsuk/times-components/compare/@times-components/article@7.4.104...@times-components/article@7.4.105) (2020-06-12)
 
 **Note:** Version bump only for package @times-components/article